requirements.txt
flask==2.3.2
python-dotenv==1.0.0
opencv-python==4.7.0.72
requests==2.31.0
Pillow==9.5.0